Updated download_URL for Lightning 1.8.6
Changes in this version: Fixed an issue where enabling FileVault on machines imaged with Lightning and macOS Sierra image could cause problems with booting the operating system. Please keep in mind that every macOS Sierra image that is used in this particular deployment scenario needs to be recreated with the newest Lightning version.